--- a/specutils/tests/test_resample.py
+++ b/specutils/tests/test_resample.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 import astropy.units as u
-from astropy.nddata import InverseVariance, StdDevUncertainty
+from astropy.nddata import VarianceUncertainty, InverseVariance, StdDevUncertainty
 from astropy.tests.helper import assert_quantity_allclose
 
 from ..spectra.spectrum1d import Spectrum1D
@@ -214,3 +214,29 @@ def test_resample_uncs(all_resamplers):
     else:
         assert resampled.uncertainty.unit == sdunc.unit
         assert resampled.uncertainty.uncertainty_type == sdunc.uncertainty_type
+
+
+def test_fluxconservingresampler_against_spectres():
+    # this test compares the output from FluxConservingResampler to the output
+    # from running the code in https://github.com/ACCarnall/SpectRes, which
+    # contains the algorithm that FluxConservedResampler is also
+    # implementing (slightly differently, but results should always agree)
+
+    spectres_fluxes = np.array([2., 3., 3.66666667, 6.3125])
+    spectres_errs = np.array([0.25, 0.15, 0.30413813, 0.49193694])
+
+    flux = [1, 2, 3, 4, 5, 6, 7]*u.AB
+    spectral_axis = [1, 2, 3, 7, 9, 10, 14]*u.AA
+    new_spec_axis = [2, 3, 4, 12]*u.AA
+    errs = VarianceUncertainty([0.5, 0.25, 0.15, 0.45, 0.75, 1.5, 0.1])
+    input_spectra = Spectrum1D(spectral_axis=spectral_axis, flux=flux,
+                               uncertainty=errs)
+
+    inst = FluxConservingResampler()
+    fluxc_resampled = inst(input_spectra, new_spec_axis)
+    fluxc_flux = fluxc_resampled.flux.value
+    # have to do 1 / errs because InverseVariance is always returned
+    fluxc_errs = 1 / fluxc_resampled.uncertainty.array
+
+    assert_quantity_allclose(fluxc_flux, spectres_fluxes)
+    assert_quantity_allclose(fluxc_errs, spectres_errs)