@pllim 2023-05-19
The timing stated in the table is very impressive. Good job! 👏 

Does @sean-lockwood want to take this PR branch for a spin with his original use case?

> I do want to note that there is a small domain where the matrix approach is faster (a million down to 2-20 samples in that case, see table)

I don't think `0.05 s -> 0.1 s` for that narrow range of samples is worth re-introducing matrix.