@cshanahan1 2023-05-26
Thanks for the review @eteq. I added a new test that takes care of item 1 on your list of suggestions - is that what you had in mind? I ran the code from the other package to generate output for the test rather than adding it as a test dependency and generating it each time.

As for items 2 and 3, I think those are good ideas and I will make follow up tickets for them but I want to get this PR merged first.