@sean-lockwood 2023-05-31
I ran the calculations again, so the numbers won't exactly match the plots above.  Values are in seconds.

```
# out      1.10.0       0.1.dev2966+g7c71c41

 200     2.145341       0.179401
 250     2.625112       0.144806
 300     3.138709       0.150052
 350     3.728590       0.147948
 400     4.217818       0.147381
 450     4.709783       0.147746
 500     5.271168       0.152145
 550     5.813444       0.150349
 600     6.293068       0.156625
 650     6.798396       0.151181
 700     7.413938       0.151452
1200    12.702604       0.157407
1700    29.881622       0.164718
2200    50.650592       0.171427
2700    73.037056       0.181266
```