@sean-lockwood 2023-05-31
Ah, I understand.  I just took a quick look at the output flux.  Most points do well, but a handful have up to 30% discrepancy in flux.  I would have to do further testing to understand the origin of the difference.

![flux_comparison](https://github.com/astropy/specutils/assets/5471441/aec4a8dd-d1f0-4e25-8512-5bd9e69a0fd4)