@cshanahan1 2023-05-31
@sean-lockwood if you can send me your input spectral axis, flux, and err values i can compare it against Spectres to make sure those differences are expected