@sean-lockwood 2023-05-31
The uncertainty array was None in my test.

Input and outputs here:
[spec_files.zip](https://github.com/astropy/specutils/files/11617745/spec_files.zip)

Read output arrays via:
```python
import pickle

with open('v1.10.0_downsampled.pkl', 'rb') as f:
    v1_wave, v1_flux = pickle.load(f)

with open('new_downsampled.pkl', 'rb') as f:
    new_wave, new_flux = pickle.load(f)
```

Generated via:
```python
import astropy.units as u
import specutils
from specutils import Spectrum1D
from specutils.manipulation import FluxConservingResampler

print(specutils.__version__)

orig = Spectrum1D.read('lte012.0-3.5-0.0a+0.0.BT-Settl.spec.fits')

resampler = FluxConservingResampler()
output_wavelengths = np.linspace(12.58e-4, 776247.11e-4, num=1500) * u.um
resampled_spec = resampler(orig, output_wavelengths)
```