@cshanahan1 2023-06-02
the bad news is that I spent too much time chasing down a non-issue, but the good news is that i determined it was a non-issue: spectres and this new version of FluxConservingResampler match basically within floating point error. (my dumb mistake and what made me worried was that i saw 0.2-0.3 offsets and interpreted these as percent differences when they were in fact absolute differences). 

it makes sense that @sean-lockwood was seeing a general agreement but an occasional 20-30% difference between the old version and the new version - those were the cases of overlapping input+output bins that were previously being treated incorrectly that i described in the linked issue. those few instances can be seen here:
<img width="807" alt="Screen Shot 2023-06-01 at 11 38 43 PM" 
src="https://github.com/astropy/specutils/assets/20444899/780504a6-37b0-44fe-99ec-fdd5598e6ac8">


what matters is that spectres and this new version agree, and they do:

<img width="797" alt="Screen Shot 2023-06-01 at 11 36 53 PM" src="https://github.com/astropy/specutils/assets/20444899/2af87f14-2319-4180-a14d-58eba62646a8">