@pllim 2023-06-02
How big is this "lte012.0-3.5-0.0a+0.0.BT-Settl.spec.fits" file? Maybe we should turn this Spectre use case into a unit test (as a follow-up PR) to prevent regression.