@sean-lockwood 2023-06-02
Attaching a truncated version of the input file (now 847k uncompressed):
[truncated_lte012.0-3.5-0.0a+0.0.BT-Settl.spec.fits.zip](https://github.com/astropy/specutils/files/11636837/truncated_lte012.0-3.5-0.0a%2B0.0.BT-Settl.spec.fits.zip)

Still gives the discrepant point:
![truncated_comparison](https://github.com/astropy/specutils/assets/5471441/51907c76-27ca-4560-8af1-40f0d3c5642e)