@cshanahan1 2023-06-02
@sean-lockwood it makes sense there are some discrepant points when comparing the new version vs the old - this PR also includes a fix for some instances where fluxes and errors were being calculated incorrectly (when input and output bins overlap in a certain way).