@pllim 2023-06-02
Thanks! I created https://github.com/astropy/specutils/issues/1064 to add a test, for whoever who wants to tackle it.